+++ integration_tests/suite/test_backend_wazo_user.py
@@ -1,11 +1,15 @@
 # Copyright 2017-2024 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+from uuid import uuid4
+
 from hamcrest import assert_that, has_entries, has_item
 from wazo_test_helpers.hamcrest.uuid_ import uuid_
 
 from .helpers import base, fixtures
 from .helpers.base import assert_http_error
+
+NEW_TENANT_UUID = str(uuid4())
 
 
 @base.use_asset('base')
@@ -14,7 +18,7 @@ class TestWazoUserBackend(base.APIIntegrationTest):
         username='foobar', email_address='foobar@example.com', password='s3cr37'
     )
     def test_token_creation(self, user):
-        response = self._post_token(user['username'], 's3cr37', backend='wazo_user')
+        response = self._post_token(user['username'], 's3cr37')
         assert_that(
             response,
             has_entries(
@@ -32,11 +36,63 @@ class TestWazoUserBackend(base.APIIntegrationTest):
             self._post_token,
             user['username'],
             'not-our-password',
-            backend='wazo_user',
         )
         assert_http_error(
-            401, self._post_token, 'not-foobar', 's3cr37', backend='wazo_user'
+            401,
+            self._post_token,
+            'not-foobar',
+            's3cr37',
         )
+
+    @fixtures.http.tenant(uuid=NEW_TENANT_UUID, default_authentication_method='saml')
+    @fixtures.http.user(
+        username='u1',
+        password='s3cr37',
+        authentication_method='ldap',
+        tenant_uuid=NEW_TENANT_UUID,
+    )
+    @fixtures.http.user(
+        username='u2',
+        password='s3cr37',
+        authentication_method='default',
+        tenant_uuid=NEW_TENANT_UUID,
+    )
+    def test_wrong_authentication_method(self, tenant, u1, u2):
+        # u1 uses ldap not native
+        assert_http_error(
+            401,
+            self._post_token,
+            'u1',
+            's3cr37',
+        )
+        # u2 uses saml (from the tenant) not native
+        assert_http_error(
+            401,
+            self._post_token,
+            'u2',
+            's3cr37',
+        )
+
+        tenant['default_authentication_method'] = 'native'
+        self.client.tenants.edit(tenant['uuid'], **tenant)
+
+        # u1 uses ldap not native
+        assert_http_error(
+            401,
+            self._post_token,
+            'u1',
+            's3cr37',
+        )
+        # u2 now uses native from the tenant
+        response = self._post_token('u2', 's3cr37')
+        assert_that(response, has_entries(token=uuid_()))
+
+        u1['authentication_method'] = 'native'
+        self.client.users.edit(u1['uuid'], **u1)
+
+        # u1 now uses native
+        response = self._post_token('u1', 's3cr37')
+        assert_that(response, has_entries(token=uuid_()))
 
     @fixtures.http.tenant()
     # extra tenant: "master" tenant
@@ -44,7 +100,7 @@ class TestWazoUserBackend(base.APIIntegrationTest):
     def test_token_metadata(self, tenant, user):
         top_tenant = self.get_top_tenant()
 
-        token_data = self._post_token(user['username'], 's3cr37', backend='wazo_user')
+        token_data = self._post_token(user['username'], 's3cr37')
 
         assert_that(
             token_data['metadata'],
@@ -60,8 +116,6 @@ class TestWazoUserBackend(base.APIIntegrationTest):
             username='foobar', email_address='foobar@example.com'
         )
         try:
-            assert_http_error(
-                401, self._post_token, user['username'], 'p45sw0rd', backend='wazo_user'
-            )
+            assert_http_error(401, self._post_token, user['username'], 'p45sw0rd')
         finally:
             self.client.users.delete(user['uuid'])